pyproject.toml
[tool.poetry]
name = "imker"
version = "0.0.1"
description = "An easy-to-use ML pipeline package for Python."
authors = ["taikinman <okipedia6161@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.25.0"
pandas = "^2.0.3"
PyYAML = "^6.0"
pydantic = "^1.10"

[tool.poetry.group.lightning]
optional = true

[tool.poetry.group.lightning.dependencies]
lightning = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
mypy = "^1.4.1"
types-PyYAML = "^6.0"
pandas-stubs = "^2.0"
ruff = "^0.0.277"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "B", # flake8-bugbear
    "I", # isort
]

ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "B024", # abstract class has no abstract methods
]

unfixable = [
    "F401", # module imported but unused
    "F841", # local variable is assigned to but never used, 
]

exclude=[
    ".git",
    "__pycache__",
    "dist", 
    "__init__.py"
]

[tool.ruff.per-file-ignores]
"example/*"=["E402"]

[tool.mypy]
python_version = 3.9
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_untyped_calls = true

[tool.black]
line-length = 100
